+++ frontend/src/components/CalendarView.vue
@@ -1,111 +1,36 @@
 <template>
   <div class="calendar__view">
-    <div
-      v-if="isMonthView"
-      class="month-view"
-    >
-      <div class="days-of-week">
-        <span
-          v-for="(day, index) in dayOfWeek"
-          :key="index"
-          class="day column"
-        >
-          {{ day }}
-        </span>
-      </div>
-      <div class="days-of-month">
-        <div
-          v-for="(day, i) in days"
-          :key="i"
-          class="day"
-          @dragover.prevent
-          @dragenter.prevent
-          @drop.prevent="onDrop(day)"
-          @click.stop="selectDay(day)"
-        >
-          <div
-            class="day-num"
-            :class="{
-              'today': isToday(day),
-              'dim': !isSameMonth(day)}"
-          >
-            {{ day.date() }}
-          </div>
-          <dl class="event-list">
-            <dt
-              v-for="(event, j) in eventsMap[day.format(dateFormat.DATE)]"
-              :key="j"
-              :class="{ 'selected': event.selected }"
-              class="event"
-              draggable="true"
-              @dragstart="onDragStart(event)"
-              @click.stop="selectEvent(event)"
-            >
-              <span>{{ event.title }}</span>
-              <span>{{ formatDate(event.start.dateTime) }}</span>
-            </dt>
-          </dl>
-        </div>
-      </div>
-    </div>
-    <table
-      v-else
-      class="week-view"
-    >
-      <thead>
-        <tr>
-          <th />
-          <th
-            v-for="(day, index) in dayOfWeek"
-            :key="index"
-          >
-            <span class="day">{{ day.format(dateFormat.DATE_AND_DAY) }}</span>
-          </th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr
-          v-for="(hour, index) in hours"
-          :key="index"
-        >
-          <td class="hour">
-            <span>{{ hour.format(dateFormat.HOUR_AND_MIN) }}</span>
-          </td>
-          <td
-            v-for="(day, i) in dayOfWeek"
-            :key="i"
-            class="hour"
-            @dragover.prevent
-            @dragenter.prevent
-            @drop.prevent="onDrop(day, hour)"
-            @click.stop="selectHour(day, hour)"
-          >
-            <div class="event-list">
-              <div
-                v-for="(event, j) in getEventsByMap(day, hour)"
-                :key="j"
-                :class="{ 'selected': event.selected }"
-                class="event"
-                draggable="true"
-                @dragstart="onDragStart(event)"
-                @click.stop="selectEvent(event)"
-              >
-                <span>{{ event.title }}</span>
-              </div>
-            </div>
-          </td>
-        </tr>
-      </tbody>
-    </table>
+    <template v-if="isMonthView">
+      <monthly
+        :current-date="currentDate"
+        :events-map="eventsMap"
+        :today="today"
+        @day-selected="selectDay"
+        @event-selected="selectEvent"
+        @event-moved="onMoveEvent"
+      />
+    </template>
+    <template v-else>
+      <weekly
+        :current-date="currentDate"
+        :events-map="eventsMap"
+        @hour-selected="selectHour"
+        @event-selected="selectEvent"
+        @event-moved="onMoveEvent"
+      />
+    </template>
   </div>
 </template>
 
 <script>
+import Monthly from '@/components/view/Monthly'
+import Weekly from '@/components/view/Weekly'
 import { viewTypes } from '@/types/calendar'
 import moment from 'moment'
 
 export default {
   name: 'CalendarView',
+  components: { Monthly, Weekly },
   props: {
     viewType: {
       type: String,
@@ -119,10 +44,6 @@ export default {
       type: Object,
       default: null
     },
-    selectedEvent: {
-      type: String,
-      default: null
-    },
     events: {
       type: Array,
       default: null
@@ -130,26 +51,12 @@ export default {
   },
   data () {
     return {
-      days: [],
-      weeks: [],
-      eventsMap: {},
-      draggingEvent: null
+      eventsMap: {}
     }
   },
   computed: {
     isMonthView () {
       return this.viewType === viewTypes.MONTH
-    },
-    dayOfWeek () {
-      if (this.viewType === viewTypes.MONTH) {
-        return moment.weekdays()
-      } else {
-        const startOfWeek = this.currentDate.clone().startOf('week').subtract(1, 'd')
-        return moment.weekdays().map(() => startOfWeek.add(1, 'd').clone())
-      }
-    },
-    hours () {
-      return Array(24).fill('').map((_, i) => moment(i, 'hh'))
     }
   },
   watch: {
@@ -158,12 +65,6 @@ export default {
         res[event.start.date] = res[event.start.date] ? [...res[event.start.date], event] : [event]
         return res
       }, {})
-    },
-    currentDate: {
-      immediate: true,
-      handler () {
-        this.buildDays()
-      }
     }
   },
   methods: {
@@ -175,43 +76,6 @@ export default {
       }
       return []
     },
-    isToday (day) {
-      return this.today.isSame(day)
-    },
-    isSameMonth (date) {
-      return this.currentDate.month() === date.month()
-    },
-    buildDays () {
-      const previous = this.daysOfPrevMonth()
-
-      const monthDate = this.currentDate.clone().startOf(viewTypes.MONTH)
-      const daysOfMonth = [...Array(monthDate.daysInMonth())].map((_, i) => monthDate.clone().add(i, 'day'))
-
-      const next = this.daysOfNextMonth()
-
-      this.days = [...previous, ...daysOfMonth, ...next]
-    },
-    daysOfPrevMonth () {
-      const startOfMonth = this.currentDate.clone().startOf('month')
-      const day = startOfMonth.day()
-      const arr = []
-      const SUNDAY_NUM = 0
-      for (let i = day; i > SUNDAY_NUM; i--) {
-        arr.push(startOfMonth.clone().subtract(i, 'd'))
-      }
-      return arr
-    },
-    daysOfNextMonth () {
-      const endOfMonth = this.currentDate.clone().endOf('month')
-      const day = endOfMonth.day()
-      const arr = []
-      const MAX_DAY = 7
-      for (let i = 1; i < MAX_DAY - day; i++) {
-        arr.push(endOfMonth.clone().add(i, 'd'))
-      }
-      return arr
-    },
-    // TODO: event delegation
     selectEvent (event) {
       this.$emit('event-selected', event)
     },
@@ -222,219 +86,9 @@ export default {
       date.hours(hour.hours())
       this.$emit('hour-selected', date)
     },
-    onDragStart (event) {
-      this.draggingEvent = event
-    },
-    onDrop (date, hour = null) {
-      const event = this.draggingEvent
-      if (event) {
-        const { start, end } = event
-        const newDate = hour ? date.clone().hours(hour.hours()) : date.clone().hours(moment(start.dateTime).hours())
-        const payload = {
-          ...event,
-          start: {
-            date: moment(start.date).month(newDate.month()).date(newDate.date()).format(this.dateFormat.DATE),
-            dateTime: moment(start.dateTime).month(newDate.month()).date(newDate.date()).hour(newDate.hours()).format(this.dateFormat.DATE_TIME)
-          },
-          end: {
-            date: moment(end.date).month(newDate.month()).date(newDate.date()).format(this.dateFormat.DATE),
-            dateTime: moment(end.dateTime).month(newDate.month()).date(newDate.date()).hour(newDate.hours() + 1).format(this.dateFormat.DATE_TIME)
-          }
-        }
-        this.$emit('event-moved', payload)
-      }
-      this.draggingEvent = null
-    },
-    formatDate (date) {
-      return moment(date).format(this.dateFormat.TIME)
+    onMoveEvent (payload) {
+      this.$emit('event-moved', payload)
     }
   }
 }
 </script>
-
-<style lang="scss" scoped>
-  $day-width: 100% / 7;
-
-  .month-view {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    .days-of-week,
-    .days-of-month {
-      display: grid;
-      grid-template-columns: repeat(7, 1fr);
-      grid-template-rows: repeat(auto-fit, minmax(50px, 1fr));
-    }
-    .days-of-month {
-      height: 100%;
-    }
-    .days-of-week {
-      height: 44px;
-      line-height: 44px;
-      background-color: var(--color-primary);
-      & > * {
-        font-size: 1.2em;
-        color: var(--color-text);
-        font-weight: 500;
-        letter-spacing: 0.1em;
-        text-align: center;
-      }
-    }
-
-    .days-of-month {
-      border-top: 1px solid var(--color-dark-gray);
-      border-left: 1px solid var(--color-dark-gray);
-      & .day {
-        position: relative;
-        display: flex;
-        flex-direction: column;
-        align-items: flex-end;
-        justify-content: flex-start;
-        overflow-x: hidden;
-        border-bottom: 1px solid var(--color-dark-gray);
-        border-right: 1px solid var(--color-dark-gray);
-        background-color: transparent;
-        color: var(--color-text);
-
-        .day-num {
-          height: 25px;
-          width: 25px;
-          border-radius: 50%;
-          display: block;
-          text-align: center;
-          line-height: 25px;
-          background-color: transparent;
-          font-weight: bold;
-          &.today {
-            background-color: var(--color-accent);
-            color: #fff;
-          }
-          &.dim {
-            color: #ccc;
-          }
-        }
-      }
-    }
-  }
-
-  .event-list {
-    width: 100%;
-    text-align: left;
-    overflow: auto;
-    .event {
-      display: flex;
-      justify-content: space-between;
-      padding: 3px;
-      cursor: pointer;
-      span {
-        &:first-child {
-          &:before {
-            flex: none;
-            margin: 0 3px 0 3px;
-            content: '';
-            display: inline-block;
-            width: 10px;
-            height: 10px;
-            -moz-border-radius: 7.5px;
-            -webkit-border-radius: 7.5px;
-            border-radius: 7.5px;
-            background-color: var(--color-info);
-          }
-        }
-        &:last-child {
-          font-size: .9em;
-        }
-      }
-      &:hover, &:focus {
-        outline: none;
-        background-color: var(--color-gray);
-        color: var(--color-text);
-      }
-      &:active {
-        color: var(--color-secondary);
-      }
-      &.selected {
-        background-color: var(--color-primary);
-      }
-    }
-  }
-
-  table.week-view {
-    width: 100%;
-    height: 100%;
-    border-spacing: 0;
-    border-collapse: separate;
-    table-layout: fixed;
-    thead {
-      tr {
-        th {
-          background: var(--color-primary);
-          color: var(--color-text);
-          padding: 0.5em;
-          overflow: hidden;
-          &:first-child {
-            width: 50px;
-            border-radius: 3px 0 0 0;
-          }
-          &:last-child {
-            border-radius: 0 3px 0 0;
-          }
-          .day {
-            font-size: 1.2em;
-            border-radius: 50%;
-            line-height: 1.8;
-            &.active {
-              background: var(--color-gray);
-              color: var(--color-primary);
-            }
-          }
-          .short {
-            display: none;
-          }
-          i {
-            vertical-align: middle;
-            font-size: 2em;
-          }
-        }
-      }
-    }
-
-    tbody {
-      tr {
-        background: var(--color-gray);
-        height: calc(100%/24);
-        &:nth-child(odd) {
-          background: var(--color-gray);
-        }
-        &:nth-child(4n+0) {
-          td {
-            border-bottom: 1px solid var(--color-primary);
-          }
-        }
-        td {
-          text-align: center;
-          vertical-align: middle;
-          border-left: 1px solid var(--color-dark-gray);
-          border-bottom: 1px solid var(--color-dark-gray);
-          position: relative;
-          height: 32px;
-          &:last-child {
-            border-right: 1px solid var(--color-dark-gray);
-          }
-          &.hour {
-            font-size: 1.2em;
-            padding: 0;
-            color: var(--color-text);
-            background: #fff;
-            border-bottom: 1px solid var(--color-dark-gray);
-            border-collapse: separate;
-            min-width: 100px;
-            span {
-              display: block;
-            }
-          }
-        }
-      }
-    }
-  }
-</style>
